fix: was broken in production feat: moved to typescript and same structure as the nuxt module example in https://v3.nuxtjs.org/guide/going-further/modules
