+++ jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   collectCoverage: true,
-  collectCoverageFrom: ['lib/**/*.js'],
+  collectCoverageFrom: ['src/**/*.js'],
   testEnvironment: 'node',
 }
