+++ tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./test/fixture/.nuxt/tsconfig.json"
+}
